requirements.txt
fastapi==0.116.1
uvicorn==0.30.6
chromadb==0.5.5
sentence-transformers==2.7.0
llama-cpp-python==0.2.90
pydantic==2.11.7
awscli==1.33.28
